requirements.txt
feedparser
requests
python-dotenv
google-generativeai
pydantic
beautifulsoup4
lxml
sentence-transformers